requirements.txt
torch~=1.7.1+cu110
torchvision~=0.8.2+cu110
numpy>=1.23.4
Pillow~=9.3.0
termcolor>=2.1.0